Stopes liking suggested sponsored posts.
I don't know if you want suggested posts to be liked, but to me it does not seem like it benefits anyone besides facebook. I want to make the option to only like pages or people but i can figure out how to navigate from the footer to get information from the header.... yet. Made some progress on the poke bomb today as well, i'm going to make a github page for it soon.  This is so exciting.